requirements.txt
fastapi
uvicorn
twilio
pymongo
pyngrok
requests
python-dotenv
python-multipart
"qrcode[pil]"
gridfs-fuse
boto3